refactor: AIモジュールの関心分離リファクタリング
## 変更内容

巨大な単一ファイル（ai-service.ts）を関心ごとに分離し、
保守性と拡張性を向上させました。

### ファイル構成

```
lib/core-modules/ai/
├── types.ts          # 型定義
├── constants.ts      # 定数（AI_SETTINGS, LOCAL_LLM_DEFAULTS）
├── index.ts          # エクスポート
├── providers/        # プロバイダ実装
│   ├── openai-provider.ts
│   ├── anthropic-provider.ts
│   └── local-provider.ts
└── services/         # サービス層
    ├── ai-service.ts # メインサービス
    └── token-utils.ts
```

### メリット

- 各プロバイダを独立してテスト・修正可能
- 新しいプロバイダの追加が容易
- 型定義・定数の再利用性向上
