+++ lib/core-modules/ai/providers/index.ts
@@ -1,0 +1,26 @@
+/**
+ * 生成AIモジュール プロバイダ層
+ *
+ * 全プロバイダを再エクスポート
+ */
+
+// Anthropic プロバイダ
+export {
+  chatWithAnthropic,
+  generateWithAnthropic,
+  translateWithAnthropic,
+} from "./anthropic-provider";
+// ローカルLLM プロバイダ
+export {
+  chatWithLocal,
+  generateWithLocal,
+  getLocalModelName,
+  testLocalConnection,
+  translateWithLocal,
+} from "./local-provider";
+// OpenAI プロバイダ
+export {
+  chatWithOpenAI,
+  generateWithOpenAI,
+  translateWithOpenAI,
+} from "./openai-provider";
