+++ lib/core-modules/ai/services/index.ts
@@ -1,0 +1,17 @@
+/**
+ * 生成AIモジュール サービス層
+ *
+ * 全サービスを再エクスポート
+ */
+
+// AIサービス
+export { AIService } from "./ai-service";
+
+// トークンユーティリティ
+export {
+  calculateContextUsage,
+  estimateMessagesTokens,
+  estimateTokens,
+  formatTokenCount,
+  getContextWindowSize,
+} from "./token-utils";
